version 2
Refactor your functionalities from part A in such a way that when a booking is made the reference is stored in the data structure and customer data is stored on a database table. Likewise, when a seat is freed the letter “F” is stored and any booking details from the database is removed.